Move all constants into single class
