+++ src/test/java/seedu/address/logic/LogicManagerTest.java
@@ -3,13 +3,13 @@ package seedu.address.logic;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static seedu.address.logic.Messages.MESSAGE_PERSON_NRIC_NOT_FOUND;
 import static seedu.address.logic.Messages.MESSAGE_UNKNOWN_COMMAND;
-import static seedu.address.logic.commands.CommandTestUtil.ADDRESS_DESC_AMY;
-import static seedu.address.logic.commands.CommandTestUtil.DOB_DESC_AMY;
-import static seedu.address.logic.commands.CommandTestUtil.EMAIL_DESC_AMY;
-import static seedu.address.logic.commands.CommandTestUtil.GENDER_DESC_AMY;
-import static seedu.address.logic.commands.CommandTestUtil.NAME_DESC_AMY;
-import static seedu.address.logic.commands.CommandTestUtil.NRIC_DESC_AMY;
-import static seedu.address.logic.commands.CommandTestUtil.PHONE_DESC_AMY;
+import static seedu.address.testutil.TestValues.ADDRESS_DESC_AMY;
+import static seedu.address.testutil.TestValues.DOB_DESC_AMY;
+import static seedu.address.testutil.TestValues.EMAIL_DESC_AMY;
+import static seedu.address.testutil.TestValues.GENDER_DESC_AMY;
+import static seedu.address.testutil.TestValues.NAME_DESC_AMY;
+import static seedu.address.testutil.TestValues.NRIC_DESC_AMY;
+import static seedu.address.testutil.TestValues.PHONE_DESC_AMY;
 import static seedu.address.testutil.Assert.assertThrows;
 import static seedu.address.testutil.TypicalPersons.AMY;
