+++ src/test/java/seedu/address/logic/commands/EditPersonDescriptorTest.java
@@ -3,13 +3,13 @@ package seedu.address.logic.commands;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static seedu.address.logic.commands.CommandTestUtil.DESC_AMY;
-import static seedu.address.logic.commands.CommandTestUtil.DESC_BOB;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_ADDRESS_BOB;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_EMAIL_BOB;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_NAME_BOB;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_PHONE_BOB;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_TAG_HUSBAND;
+import static seedu.address.testutil.TestValues.DESC_AMY;
+import static seedu.address.testutil.TestValues.DESC_BOB;
+import static seedu.address.testutil.TestValues.VALID_ADDRESS_BOB;
+import static seedu.address.testutil.TestValues.VALID_EMAIL_BOB;
+import static seedu.address.testutil.TestValues.VALID_NAME_BOB;
+import static seedu.address.testutil.TestValues.VALID_PHONE_BOB;
+import static seedu.address.testutil.TestValues.VALID_TAG_HUSBAND;
 
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
