+++ src/test/java/seedu/address/logic/parser/AddMedConCommandParserTest.java
@@ -2,17 +2,17 @@ package seedu.address.logic.parser;
 
 import static seedu.address.logic.Messages.MESSAGE_CONSTRAINTS_LENGTH;
 import static seedu.address.logic.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
-import static seedu.address.logic.commands.CommandTestUtil.INVALID_MEDCON_DESC;
-import static seedu.address.logic.commands.CommandTestUtil.INVALID_NRIC_DESC;
-import static seedu.address.logic.commands.CommandTestUtil.MEDCON_DESC_AMY;
-import static seedu.address.logic.commands.CommandTestUtil.MEDCON_DESC_BOB;
-import static seedu.address.logic.commands.CommandTestUtil.NAME_DESC_AMY;
-import static seedu.address.logic.commands.CommandTestUtil.NRIC_DESC_AMY;
-import static seedu.address.logic.commands.CommandTestUtil.NRIC_DESC_BOB;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_MEDCON_AMY;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_MEDCON_BOB;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_NRIC_AMY;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_NRIC_BOB;
+import static seedu.address.testutil.TestValues.INVALID_MEDCON_DESC;
+import static seedu.address.testutil.TestValues.INVALID_NRIC_DESC;
+import static seedu.address.testutil.TestValues.MEDCON_DESC_AMY;
+import static seedu.address.testutil.TestValues.MEDCON_DESC_BOB;
+import static seedu.address.testutil.TestValues.NAME_DESC_AMY;
+import static seedu.address.testutil.TestValues.NRIC_DESC_AMY;
+import static seedu.address.testutil.TestValues.NRIC_DESC_BOB;
+import static seedu.address.testutil.TestValues.VALID_MEDCON_AMY;
+import static seedu.address.testutil.TestValues.VALID_MEDCON_BOB;
+import static seedu.address.testutil.TestValues.VALID_NRIC_AMY;
+import static seedu.address.testutil.TestValues.VALID_NRIC_BOB;
 import static seedu.address.logic.parser.CommandParserTestUtil.assertParseFailure;
 import static seedu.address.logic.parser.CommandParserTestUtil.assertParseSuccess;
