+++ src/test/java/seedu/address/logic/parser/DeleteApptCommandParserTest.java
@@ -1,23 +1,23 @@
 package seedu.address.logic.parser;
 
 import static seedu.address.logic.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
-import static seedu.address.logic.commands.CommandTestUtil.APPOINTMENT_DESC_AMY;
-import static seedu.address.logic.commands.CommandTestUtil.APPOINTMENT_DESC_BOB;
-import static seedu.address.logic.commands.CommandTestUtil.INVALID_DEL_APPT_DATE_FORMAT_DESC;
-import static seedu.address.logic.commands.CommandTestUtil.INVALID_DEL_APPT_DATE_NONEXISTANT_DESC;
-import static seedu.address.logic.commands.CommandTestUtil.INVALID_DEL_APPT_TIMEPERIOD_FORMAT_DESC;
-import static seedu.address.logic.commands.CommandTestUtil.INVALID_DEL_APPT_TIMEPERIOD_ORDER_DESC;
-import static seedu.address.logic.commands.CommandTestUtil.INVALID_NRIC_DESC;
-import static seedu.address.logic.commands.CommandTestUtil.NRIC_DESC_AMY;
-import static seedu.address.logic.commands.CommandTestUtil.NRIC_DESC_BOB;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_APPOINTMENT_DATE_AMY;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_APPOINTMENT_DATE_BOB;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_APPOINTMENT_DATE_DENTAL;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_APPOINTMENT_TIMEPERIOD_AMY;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_APPOINTMENT_TIMEPERIOD_BOB;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_APPOINTMENT_TIMEPERIOD_DENTAL;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_NRIC_AMY;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_NRIC_BOB;
+import static seedu.address.testutil.TestValues.APPOINTMENT_DESC_AMY;
+import static seedu.address.testutil.TestValues.APPOINTMENT_DESC_BOB;
+import static seedu.address.testutil.TestValues.INVALID_DEL_APPT_DATE_FORMAT_DESC;
+import static seedu.address.testutil.TestValues.INVALID_DEL_APPT_DATE_NONEXISTANT_DESC;
+import static seedu.address.testutil.TestValues.INVALID_DEL_APPT_TIMEPERIOD_FORMAT_DESC;
+import static seedu.address.testutil.TestValues.INVALID_DEL_APPT_TIMEPERIOD_ORDER_DESC;
+import static seedu.address.testutil.TestValues.INVALID_NRIC_DESC;
+import static seedu.address.testutil.TestValues.NRIC_DESC_AMY;
+import static seedu.address.testutil.TestValues.NRIC_DESC_BOB;
+import static seedu.address.testutil.TestValues.VALID_APPOINTMENT_DATE_AMY;
+import static seedu.address.testutil.TestValues.VALID_APPOINTMENT_DATE_BOB;
+import static seedu.address.testutil.TestValues.VALID_APPOINTMENT_DATE_DENTAL;
+import static seedu.address.testutil.TestValues.VALID_APPOINTMENT_TIMEPERIOD_AMY;
+import static seedu.address.testutil.TestValues.VALID_APPOINTMENT_TIMEPERIOD_BOB;
+import static seedu.address.testutil.TestValues.VALID_APPOINTMENT_TIMEPERIOD_DENTAL;
+import static seedu.address.testutil.TestValues.VALID_NRIC_AMY;
+import static seedu.address.testutil.TestValues.VALID_NRIC_BOB;
 import static seedu.address.logic.commands.DeleteApptCommand.MESSAGE_USAGE;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_DATE;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_TIMEPERIOD;
