+++ src/test/java/seedu/address/model/appointment/AppointmentTest.java
@@ -1,13 +1,13 @@
 package seedu.address.model.appointment;
 
-import static seedu.address.logic.commands.CommandTestUtil.INVALID_APPOINTMENT_DATE_FORMAT;
-import static seedu.address.logic.commands.CommandTestUtil.INVALID_APPOINTMENT_DATE_NONEXISTANT;
-import static seedu.address.logic.commands.CommandTestUtil.INVALID_APPOINTMENT_NAME;
-import static seedu.address.logic.commands.CommandTestUtil.INVALID_APPOINTMENT_TIMEPERIOD_FORMAT;
-import static seedu.address.logic.commands.CommandTestUtil.INVALID_APPOINTMENT_TIMEPERIOD_ORDER;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_APPOINTMENT_DATE_DENTAL;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_APPOINTMENT_NAME_DENTAL;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_APPOINTMENT_TIMEPERIOD_DENTAL;
+import static seedu.address.testutil.TestValues.INVALID_APPOINTMENT_DATE_FORMAT;
+import static seedu.address.testutil.TestValues.INVALID_APPOINTMENT_DATE_NONEXISTANT;
+import static seedu.address.testutil.TestValues.INVALID_APPOINTMENT_NAME;
+import static seedu.address.testutil.TestValues.INVALID_APPOINTMENT_TIMEPERIOD_FORMAT;
+import static seedu.address.testutil.TestValues.INVALID_APPOINTMENT_TIMEPERIOD_ORDER;
+import static seedu.address.testutil.TestValues.VALID_APPOINTMENT_DATE_DENTAL;
+import static seedu.address.testutil.TestValues.VALID_APPOINTMENT_NAME_DENTAL;
+import static seedu.address.testutil.TestValues.VALID_APPOINTMENT_TIMEPERIOD_DENTAL;
 import static seedu.address.testutil.Assert.assertThrows;
 
 import org.junit.jupiter.api.Assertions;
