+++ src/test/java/seedu/address/testutil/TypicalPersons.java
@@ -1,21 +1,21 @@
 package seedu.address.testutil;
 
-import static seedu.address.logic.commands.CommandTestUtil.VALID_ADDRESS_AMY;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_ADDRESS_BOB;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_DOB_AMY;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_DOB_BOB;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_EMAIL_AMY;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_EMAIL_BOB;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_GENDER_AMY;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_GENDER_BOB;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_NAME_AMY;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_NAME_BOB;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_NRIC_AMY;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_NRIC_BOB;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_PHONE_AMY;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_PHONE_BOB;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_TAG_FRIEND;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_TAG_HUSBAND;
+import static seedu.address.testutil.TestValues.VALID_ADDRESS_AMY;
+import static seedu.address.testutil.TestValues.VALID_ADDRESS_BOB;
+import static seedu.address.testutil.TestValues.VALID_DOB_AMY;
+import static seedu.address.testutil.TestValues.VALID_DOB_BOB;
+import static seedu.address.testutil.TestValues.VALID_EMAIL_AMY;
+import static seedu.address.testutil.TestValues.VALID_EMAIL_BOB;
+import static seedu.address.testutil.TestValues.VALID_GENDER_AMY;
+import static seedu.address.testutil.TestValues.VALID_GENDER_BOB;
+import static seedu.address.testutil.TestValues.VALID_NAME_AMY;
+import static seedu.address.testutil.TestValues.VALID_NAME_BOB;
+import static seedu.address.testutil.TestValues.VALID_NRIC_AMY;
+import static seedu.address.testutil.TestValues.VALID_NRIC_BOB;
+import static seedu.address.testutil.TestValues.VALID_PHONE_AMY;
+import static seedu.address.testutil.TestValues.VALID_PHONE_BOB;
+import static seedu.address.testutil.TestValues.VALID_TAG_FRIEND;
+import static seedu.address.testutil.TestValues.VALID_TAG_HUSBAND;
 
 import java.util.ArrayList;
 import java.util.Arrays;
